pyproject.toml
[tool.poetry]
name = "repl_python3_beta_mogus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-game-info = "^0.1.3"
Flask = "^2.0.2"
replit = "^3.2.4"
discord-components = "^2.1.2"
buttons = "^0.1.9"
pytz = "^2021.3"
repl = "^1.0"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
